+++ Packages/com.nickmaltbie.MinimapTools/Tests/PlayMode/SamplePlaymodeTests.cs
@@ -16,38 +16,18 @@
 // ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using System.Collections;
-using com.nickmaltbie.MinimapTools;
+using nickmaltbie.MinimapTools.TestCommon;
 using NUnit.Framework;
-using UnityEngine;
-using UnityEngine.TestTools;
 
 namespace nickmaltbie.MinimapTools.Tests.PlayMode
 {
-    /// <summary>
-    /// Simple tests meant to be run in PlayMode
-    /// </summary>
-    public class PlayModeTests
+    [TestFixture]
+    public class SamplePlaymodeTests : TestBase
     {
-        /// <summary>
-        /// Simple sample script test.
-        /// </summary>
-        [UnityTest]
-        public IEnumerator SimpleSampleScriptTest()
+        [Test]
+        public void Sample_PlaymodeTest()
         {
-            var go = new GameObject();
-            SampleScript sample = go.AddComponent<SampleScript>();
-
-            Assert.AreEqual(sample.Value, 0);
-
-            yield return null;
-            Assert.IsTrue(sample.Value > 0);
-            int value = sample.Value;
-
-            yield return new WaitForSeconds(1.0f);
-            Assert.IsTrue(sample.Value > value);
-
-            GameObject.DestroyImmediate(go);
+            Assert.IsTrue(true);
         }
     }
 }