+++ Packages/com.nickmaltbie.MinimapTools/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## In Progress
 
+* Setup simple static minimap and project configuration.
 * Added proposed design docs for project.
 * Setup basic documentation for project.
 * Renamed project to Minimap Tools.